Extract ShapeRenderer interface and per-classification renderers
ModelGenerator is now a thin dispatcher that routes classified shapes to
registered ShapeRenderer implementations. Adding new classification types
(e.g., event streams) requires only a new renderer class, not editing
the orchestrator.

- EnumShapeRenderer, SubObjectRenderer, RequestRenderer, ResultRenderer
- Consolidate isPrimitive() into CppTypeMapper
- Promote writeDocComment/collapseWhitespace to public in MemberRenderer
- No behavioral change to generated output
